+++ packages/yak_result/lib/src/extensions/all.dart
@@ -3,3 +3,5 @@ export 'as_failure.dart';
 export 'failure_recast.dart';
 export 'as_void.dart';
 export 'as_success.dart';
+export 'iterable_combine.dart';
+export 'iterable_combine_unary.dart';