+++ packages/yak_runner/lib/src/extension/all.dart
@@ -2,3 +2,4 @@ export 'let_run.dart';
 export 'run_delegate.dart';
 export 'run.dart';
 export 'then_run.dart';
+export 'iterable_run.dart';